+++ Source/Jobs/Controllers/JobOrderInspectionRequestSettingsController.cs
@@ -106,7 +106,7 @@ namespace Jobs.Controllers
 
                     LogActivity.LogActivityDetail(LogVm.Map(new ActiivtyLogViewModel
                     {
-                        DocTypeId = pt.DocTypeId,
+                        DocTypeId = (int)pt.DocTypeId,
                         DocId = pt.JobOrderInspectionRequestSettingsId,
                         ActivityType = (int)ActivityTypeContants.SettingsAdded,
                     }));
@@ -160,7 +160,7 @@ namespace Jobs.Controllers
 
                     LogActivity.LogActivityDetail(LogVm.Map(new ActiivtyLogViewModel
                     {
-                        DocTypeId = temp.DocTypeId,
+                        DocTypeId = (int)temp.DocTypeId,
                         DocId = temp.JobOrderInspectionRequestSettingsId,
                         ActivityType = (int)ActivityTypeContants.SettingsModified,
                         xEModifications = Modifications,
@@ -237,7 +237,7 @@ namespace Jobs.Controllers
 
                     LogActivity.LogActivityDetail(LogVm.Map(new ActiivtyLogViewModel
                     {
-                        DocTypeId = pt.DocTypeId,
+                        DocTypeId = (int)pt.DocTypeId,
                         DocId = pt.JobOrderInspectionRequestSettingsId,
                         ActivityType = (int)ActivityTypeContants.SettingsAdded,
                     }));
@@ -291,7 +291,7 @@ namespace Jobs.Controllers
 
                     LogActivity.LogActivityDetail(LogVm.Map(new ActiivtyLogViewModel
                     {
-                        DocTypeId = temp.DocTypeId,
+                        DocTypeId = (int)temp.DocTypeId,
                         DocId = temp.JobOrderInspectionRequestSettingsId,
                         ActivityType = (int)ActivityTypeContants.SettingsModified,
                         xEModifications = Modifications,
